refactor(db): extract focused repository types from db.DB (#506)
- Add ScanRepository, HostRepository (extended), ProfileRepository,
  ScheduleRepository, DiscoveryRepository in repository_*.go files
- Each repository wraps *DB and implements its matching handler store
  interface; ScanRepository.GetProfile delegates to ProfileRepository
- Remove all entity methods from *DB — it now holds only Connect(),
  Close(), Ping(), BeginTx(), and shared query helpers
- Update routes.go and handlers/manager.go to wire repositories
- Update cmd/cli/profiles.go to use db.NewProfileRepository(database)
- Update all unit/integration test files to call through repository
  constructors instead of *DB methods directly
- Fix lll violations in new files; fix goconst (use ScanJobStatusCompleted)
